[header file] Update header file to fix warning
- Update included header file to fix warning
  to avoid potential issue in mx8dv

+++ tools/grecorder/fsl_types.h
@@ -38,10 +38,10 @@
 #endif
 #endif /*int64*/
 
-typedef unsigned long uint32;
+typedef unsigned int uint32;
 typedef unsigned short uint16;
 typedef unsigned char uint8;
-typedef long int32;
+typedef int int32;
 typedef short int16;
 typedef char int8;
